Module Relations.XOR_Rotate

type (_, _) t = private

| XR_Identity : ('a, 'a) t

| XR_BNot : (Operator.Function_symbol.boolean, Operator.Function_symbol.boolean) t
(*
Boolean not
*)

| XR_XOR_rotate : {

rotate : int;

xor : Z.t;

size : Units.In_bits.t;

} -> (Operator.Function_symbol.bitvector, Operator.Function_symbol.bitvector) t
(*
On bounded integers, represents a rotation then a xor Invariant: xor fits on size bits, 0 <= rotate < size
*)

Bitwise focused relation: xor and rotate.

val xr_identity : ('a, 'a) t

val xr_bnot : (Operator.Function_symbol.boolean, Operator.Function_symbol.boolean) t

val xr_xor_rotate : rotate:int -> xor:Z.t -> size:Units.In_bits.t -> (Operator.Function_symbol.bitvector, Operator.Function_symbol.bitvector) t

include Union_Find.Parameters.GENERIC_GROUP with type ('a, 'b) t := ('a, 'b) t

include Union_Find.Parameters.GENERIC_MONOID with type ('a, 'b) t := ('a, 'b) t

val equal : ('a, 'b) t -> ('a, 'b) t -> bool

Equality of relations

val pretty : Format.formatter -> ('a, 'b) t -> unit

Pretty printer for relations

val pretty_with_terms : (Format.formatter -> 'tl -> unit) -> 'tl -> (Format.formatter -> 'tr -> unit) -> 'tr -> Format.formatter -> ('a, 'b) t -> unit

pretty_with_terms pp_x x pp_y y rel pretty-prints the relation rel between terms x and y (respectively printed with pp_x and pp_y).

For placeholder variables, use pretty

val identity : ('a, 'a) t

The identity relation

val compose : ('b, 'c) t -> ('a, 'b) t -> ('a, 'c) t

Monoid composition, written using the functional convention compose f g is f \circ g. Should be associative, and compatible with identity:

For all x, G.compose x G.identity = G.compose G.identity x = x

For all x y z, G.compose x (G.compose y z) = G.compose (G.compose x y) z

val inverse : ('a, 'b) t -> ('b, 'a) t

Group inversion, should verify for all x: G.compose x (G.inverse x) = G.compose (G.inverse x) x = G.identity