Module Slicing.Make

Parameters

module T : Sig.TERMS

Signature

val deps : 'a T.t -> T.cfg_node -> int list

Given a term variable x, compute for each tuple argument in each CFG node, on which x depends, the set of indices in the tuples on which x depends.

Note that dependency computation is easy (it is just the transitive closure of the points-to relation in the term graph), and only requires to do a first pass to get these indices.